### 2021-10-14 #3825 @whedon
[PRE REVIEW]: Sepia, a taxonomy oriented read classifier in Rust
**Submitting author:** @hcdenbakker (<a href="http://orcid.org/0000-0002-4086-1580">Henk den Bakker</a>)
**Repository:** <a href="https://github.com/hcdenbakker/sepia" target ="_blank">https://github.com/hcdenbakker/sepia</a>
**Version:** v0.0.2
