### 2021-10-21 hcdenbakker/sepia#13 @Ebedthan
Fix cargo clippy warnings
Hi @hcdenbakker,

I believe that one core advantage of Rust is the detection of potential threats to code health at compile time. So I recommend you to follow the `cargo clippy` recommendation and fix the warning pointed out.  So you could remove the "ignore warning" in the installation instruction.
Please do it as soon as possible to let the review process go forward as I will point this issue in my review report.
