### 2021-10-21 aa84cc9
Add help string for --index
