+++ test.sh
@@ -1,45 +1,115 @@
-#compile sepia in release mode
-echo "compile sepia"
-cargo build --release
-if [ $? -gt 0 ]; then
-  echo "ERROR compiling sepia, did you install Rust(https://www.rust-lang.org/)?";
-  exit 1
+#!/bin/bash
+
+# This script directory
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+# function "build" to compile
+function build() {
+  cargo build --release
+  if [ $? -gt 0 ]; then
+        echo "ERROR compiling sepia, did you install Rust(https://www.rust-lang.org/)?";
+        exit 1
+  fi
+}
+
+function print_separator() {
+  GREEN_LINE="\033[0;32m -------------------------------------------------------------- \033[0m"
+  echo -e "${GREEN_LINE}"
+}
+
+# DEFAULTS
+THREADS=2
+
+# Get options: -b build, -c clean
+while getopts "bcht:" opt; do
+  case $opt in
+    # BUILD
+    b)
+      echo " [*] Building"
+      build
+      ;;
+    # CLEAN
+    c)
+      echo " [*] Cleaning..."
+      for FILE in target/release demo_index {test_classify,Ecoli,Efaecium}*.{gz,txt};
+      do
+        
+        if [[ -e "$DIR"/"$FILE" ]]; then
+          rm -rf "$DIR"/"$FILE"
+        fi
+      done
+      exit
+      ;;
+    # -t INT (Threads)
+    t)
+      echo " - Setting THREADS=$OPTARG"
+      THREADS=$OPTARG
+      ;;
+    h)
+      echo "Help:"
+      echo "  -b:     Force rebuilding the binary"
+      echo "  -c:     Clean test.sh products (binaries and output files)"
+      echo "  -t INT: Number of threads for multithreading [default: 2]"
+      echo "  -h:     help"
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG"
+      ;;
+    *)
+      echo "Invalid option: -$OPTARG"
+      ;;
+  esac
+done
+ 
+
+
+#compile sepia in release mode if the binary is not already present
+
+if [[ ! -e "$DIR"/target/release/sepia ]]; then
+  print_separator
+  echo "Sepia binary not found...";
+  build
 fi
 
 # Build the database with one thread
-echo "Building the index"
+print_separator;
+echo " [1] Building the index";
 ./target/release/sepia build -k 31 -m 21 -i ./demo_index -r ./ref_demo.txt
-if [ $? -gt 0 ]; then
+if [[ $? -gt 0 ]]; then
   echo "ERROR building sepia index ./demo_index using ./ref_demo.txt";
   exit 1
 fi
 
 # Build the database with two threads
-echo "Building the database"
-./target/release/sepia build -k 31 -m 21 -i ./demo_index -r ./ref_demo.txt -p 2 -c 2
-if [ $? -gt 0 ]; then
+print_separator
+echo " [2] Building the index (multithreads)"
+./target/release/sepia build -k 31 -m 21 -i ./demo_index -r ./ref_demo.txt -p $THREADS -c 2
+if [[ $? -gt 0 ]]; then
   echo "ERROR building sepia index ./demo_index using ./ref_demo.txt";
   exit 1
 fi
 
 #single end read classification
-echo "Classifying single end reads"
-./target/release/sepia classify -i ./demo_index -q ./1000K_test_fq/Ecoli1_1000K_1.fastq.gz -n test_classify -t 4
+print_separator
+echo " [3] Classifying single end reads"
[…]
