### 2021-10-21 597ba05
Update test script with clean and build flags
