### 2021-10-21 bf25713
fix warnings at compile time
