### 2021-10-21 b97810e
Keep 'paper' folder out of the main branch
