+++ paper/paper.bib
@@ -1,78 +1,0 @@
-@article{wood2019improved,
-  title={Improved metagenomic analysis with Kraken 2},
-  author={Wood, Derrick E and Lu, Jennifer and Langmead, Ben},
-  journal={Genome biology},
-  volume={20},
-  number={1},
-  pages={1--13},
-  year={2019},
-  publisher={Springer}
-}
-
-@article{wood2014kraken,
-  title={Kraken: ultrafast metagenomic sequence classification using exact alignments},
-  author={Wood, Derrick E and Salzberg, Steven L},
-  journal={Genome biology},
-  volume={15},
-  number={3},
-  pages={1--12},
-  year={2014},
-  publisher={BioMed Central}
-}
-
-@article{kim2016centrifuge,
-  title={Centrifuge: rapid and sensitive classification of metagenomic sequences},
-  author={Kim, Daehwan and Song, Li and Breitwieser, Florian P and Salzberg, Steven L},
-  journal={Genome research},
-  volume={26},
-  number={12},
-  pages={1721--1729},
-  year={2016},
-  publisher={Cold Spring Harbor Lab}
-}
-
-@article{limasset2017fast,
-  title={Fast and scalable minimal perfect hashing for massive key sets},
-  author={Limasset, Antoine and Rizk, Guillaume and Chikhi, Rayan and Peterlongo, Pierre},
-  journal={arXiv preprint arXiv:1702.03154},
-  year={2017}
-}
-
-@article{parks2018standardized,
-  title={A standardized bacterial taxonomy based on genome phylogeny substantially revises the tree of life},
-  author={Parks, Donovan H and Chuvochina, Maria and Waite, David W and Rinke, Christian and Skarshewski, Adam and Chaumeil, Pierre-Alain and Hugenholtz, Philip},
-  journal={Nature biotechnology},
-  volume={36},
-  number={10},
-  pages={996--1004},
-  year={2018},
-  publisher={Nature Publishing Group}
-}
-
-@misc{appleby2011murmurhash3,
-  title={MurmurHash3 Hash Function},
-  author={Appleby, A},
-  year={2011}
-}
-
-@article{nayfach2016integrated,
-  title={An integrated metagenomics pipeline for strain profiling reveals novel patterns of bacterial transmission and biogeography},
-  author={Nayfach, Stephen and Rodriguez-Mueller, Beltran and Garud, Nandita and Pollard, Katherine S},
-  journal={Genome research},
-  volume={26},
-  number={11},
-  pages={1612--1625},
-  year={2016},
-  publisher={Cold Spring Harbor Lab}
-}
-
-@article{segata2012metagenomic,
-  title={Metagenomic microbial community profiling using unique clade-specific marker genes},
-  author={Segata, Nicola and Waldron, Levi and Ballarini, Annalisa and Narasimhan, Vagheesh and Jousson, Olivier and Huttenhower, Curtis},
-  journal={Nature methods},
-  volume={9},
-  number={8},
-  pages={811--814},
-  year={2012},
-  publisher={Nature Publishing Group}
-}
