+++ paper/paper.md
@@ -1,56 +1,0 @@
----
-title: ‘Sepia, a taxonomy oriented read classifier in Rust’
-authors:
-- affiliation: 1
-  name: Henk C. den Bakker^[corresponding author]
-  orcid: 0000-0002-4086-1580
-- affiliation: "1, 2"
-  name: Lee S. Katz
-  orcid: 0000-0002-2533-9161
-date: “14 October, 2021”
-output:
-  html_document:
-    df_print: paged
-  pdf_document: default
-  word_document: default
-bibliography: paper.bib
-tags:
-- read classification
-- taxonomy
-affiliations:
-- index: 1
-  name: Center for Food Safety, University of Georgia, Griffin, GA, USA
-- index: 2
-  name: Enteric Diseases Laboratory Branch (EDLB), Centers for Disease Control and Prevention, Atlanta, GA, USA
----
- 
-# Summary
-
-Here we present Sepia, a fast and accurate read classifier. It is implemented in Rust, has the ability to switch between different taxonomies, can detect inconsistencies in taxonomies, and can estimate similarities between organisms in a query dataset and the index.
-
-# Statement of need
-
-Bioinformatics tools to infer the taxonomic composition of sets of biological sequences are quintessential for taxonomic profiling and contamination checking. A variety of tools have been developed to accomplish taxonomic profiling and these tools can be roughly classified into two categories: (i) read alignment based tools such as MIDAS [@nayfach2016integrated] and MetaPhlan [@segata2012metagenomic] , which map reads against a database of reference sequences or taxon specific marker sequences, and (ii) read classification based tools, such as Kraken [@wood2014kraken], Kraken2 [@wood2019improved] and Centrifuge [@kim2016centrifuge]. Many read classifiers use a least common ancestor (LCA) approach where reads in common among sibling taxa get assigned to the rank and taxonomy in common with those taxa; e.g., a kmer in common with both Escherichia and Salmonella, both members of the family Enterobacteriaceae, would be elevated to be associated in a database with the higher order taxon Enterobacteriaceae instead. While all these tools heavily rely on taxonomies, changing the taxonomies (e.g., correcting wrong placements of accessions or adapting a novel taxonomy) is not easily done.  In response, we wrote Sepia, a taxonomic read classifier to address rapidly changing developments in taxonomy (e.g., the genome-based GTDB taxonomy [@parks2018standardized]) and algorithm development. There are three areas where Sepia directly addresses issues needing improvement: (i) taxonomy, (ii) classification accuracy and (iii) the ability to perform fast batch classification for multiple datasets.
-
-## Taxonomy
-Taxonomic read classifiers that use an LCA approach work on the assumption that biological taxonomies reflect evolutionary relationships, thus making it possible to use taxonomies as a predictive framework in read classification. Inclusion of artificial taxa that are not supported by a genome-based phylogeny (e.g., garbage bin taxa or taxa with an unknown placement ‘incertae sedis’) or artifactual errors (e.g., a genus placed erroneously in the wrong family in some of but not all of the accessions in the taxonomy) have a deleterious effect on the accuracy LCA based classification algorithms, potentially leading to some taxa not being classified at all or consistently classified as the wrong taxon. 
-To address these issues Sepia uses a human-readable taxonomy strings as input. While building the index putative ambiguities or inconsistencies in the taxonomy are flagged and logged to a file for the user to address.
-
-## Sequence similarity and classification accuracy
[…]
