### 2021-10-26 51e390f
first round of clippy edits
