### 2021-11-02 hcdenbakker/sepia#13 @Ebedthan
Hi @hcdenbakker,
Please can you kindly let me know if you are done with fixing Clippy warnings?
Although I can agree with you in some cases, `cargo clippy` actually points some issues that I believe _should_ be resolved for the general health of the application.
For example, Clippy points to a ``use of `unwrap_or` followed by a function call`` which is an [unnecessary lazy evaluation](https://rust-lang.github.io/rust-clippy/master/index.html#or_fun_call).
You will find many other warnings that I advise you to fix as I believe it will not have a negative impact on your application but moreover make your application safer.
