### 2021-11-02 #3839 @will-rowe
Hi all. Just checking in. Thanks @telatin for your thorough review. @hcdenbakker - can you please respond to the minor comments over in https://github.com/hcdenbakker/sepia/issues/17

@Ebedthan - how are you getting on with the review?
