### 2021-11-02 #3839 @Ebedthan
Dear @hcdenbakker,

Thank you for the interesting submission. This software fills a gap in the taxonomic read tools.
I have been able to install, test, and run the software without any issue.

Overall, **the repository** is well put together. However, I have some comments which I think will make the software more accessible to the community. I have listed them below with links to the corresponding issues I have opened on the software repository. I will also be happy to discuss any of the points further if clarification is needed.

* I believe the code will benefit ([issue here](https://github.com/hcdenbakker/sepia/issues/13)) from incorporating some advises as suggested by `cargo clippy`.

* The README is well written however it might benefit from a clearer installation section. The installation section should be divided into "precompiled binaries" and "install from source" sections. The release pages should be linked in the precompiled binaries section. The example code in the "install from source" section would be better on new lines like below:

> 2. Clone the sepia repository
> ```
> git clone https://github.com/hcdenbakker/sepia.git
> ```
* The "Running sepia" section can be clearer by beginning with the "Build an index section" followed with "Classifying reads" and "Batch mode" thus showing the typical pipeline a user would follow. I think that the global sepia help page can come at the end of this section.

* Using `cargo test` I've found only one unit test for the entire software. I recommend adding more tests and making the coverage of the test suite available in the README.  Ideally, you should also set up continuous integration as
well which can be done with GitHub actions. Although I do not believe this is a specific requirement for JOSS I would advise it. You can find some examples on how to do it with bio crates like [yacrd](https://github.com/natir/yacrd) and [sabreur](https://github.com/Ebedthan/sabreur).
* The To do at the end of the README can be placed inside the main.rs as a comment.
* The precompiled binaries are marked with a pre-release badge. If it is not intended, I would recommend removing it.
* I would suggest the release to have a proper naming strategy.
* The precompiled binaries should contain the target triple.

In terms of **the paper**:
* Line 8: reword "the index"
* Line 26: add a comma before "and"
* Line 36-37: the indefinite article "a" may not be required with the plural noun "strings" in this sentence.

I hope you find these comments useful! As I have said I have overall found the tools helpful and well written.

Best,
Anicet.
