+++ src/search_bits.rs
@@ -2,6 +2,7 @@ use super::bit_magic::get_phf;
 use super::kmer;
 use boomphf::*;
 use needletail::parse_fastx_file;
+use super::build_index::Parameters;
 
 use flate2::write::GzEncoder;
 use flate2::Compression;
@@ -62,8 +63,7 @@ fn sliding_window_minimizers_phf_vec(
                     window.pop_back(); // we pop the last one
                 }
                 window.push_back((minimizer, i)); // and make add a pair with the new value at the end
-                while window.front().unwrap().1 as isize <= i as isize - k as isize + m as isize - 1
-                {
+                while (window.front().unwrap().1 as isize) < i as isize - k as isize + m as isize {
                     window.pop_front(); // pop the first one
                 }
                 if i >= k {
@@ -73,7 +73,7 @@ fn sliding_window_minimizers_phf_vec(
                         *count += 1;
                     } else {
                         //continue
-                        taxon = get_phf(&(window.front().unwrap().0 ^ toggle), &db, phf, b);
+                        taxon = get_phf(&(window.front().unwrap().0 ^ toggle), db, phf, b);
                         let count = report.entry(taxon).or_insert(0);
                         *count += 1;
                         current_minimizer = window.front().unwrap().0;
@@ -102,7 +102,7 @@ fn sliding_window_minimizers_phf_vec(
 pub fn search_index_lca(map: &[String], db: &[u32], bloom_size: u64) -> HashMap<u32, usize> {
     let mut report = HashMap::default();
     for k in map {
-        let position = seahash::hash(&k.as_bytes()) % bloom_size;
+        let position = seahash::hash(k.as_bytes()) % bloom_size;
         let count = report.entry(db[position as usize]).or_insert(0);
         *count += 1;
     }
@@ -118,23 +118,21 @@ pub fn score_lineages(
     let mut lineage_scores = HashMap::default();
     let unclassified = (taxonomy.len() + 1) as u32;
     for key in report.keys() {
-        if *key == (0 as u32) || *key == unclassified {
+        if *key == (0_u32) || *key == unclassified {
             continue;
         } else {
-            let lineage1 = &taxonomy[&key];
+            let lineage1 = &taxonomy[key];
             for key2 in report.keys() {
-                if *key2 == (0 as u32) || *key2 == unclassified {
+                if *key2 == (0_u32) || *key2 == unclassified {
                     continue;
                 } else {
-                    let lineage2 = &taxonomy[&key2];
+                    let lineage2 = &taxonomy[key2];
                     if lineage1 == lineage2 {
                         let count = lineage_scores.entry(*key).or_insert(0);
-                        *count += &report[&key];
-                    } else {
-                        if lineage1.contains(lineage2) {
-                            let count = lineage_scores.entry(*key).or_insert(0);
-                            *count += &report[&key2];
-                        }
+                        *count += &report[key];
+                    } else if lineage1.contains(lineage2) {
+                        let count = lineage_scores.entry(*key).or_insert(0);
+                        *count += &report[key2];
                     }
                 }
             }
@@ -155,20 +153,20 @@ pub fn poll_lineages(
     let unclassified = (taxonomy.len() + 1) as u32;
     for key in report.keys() {
         let mut score = 0;
-        if *key == (0 as u32) || *key == unclassified {
+        if *key == (0_u32) || *key == unclassified {
             continue;
         } else {
[…]
