+++ src/kmer.rs
@@ -24,7 +24,7 @@ pub fn read_fasta(filename: String) -> Vec<String> {
         count_line += 1;
         if line.starts_with('>') {
             let l = sub_string.to_string();
-            if l.len() > 0 {
+            if !l.is_empty() {
                 vec.push(l);
             }
             sub_string.clear();
@@ -32,7 +32,7 @@ pub fn read_fasta(filename: String) -> Vec<String> {
             let l = line.to_string();
             sub_string.push_str(&l);
             let l = sub_string.to_string();
-            if l.len() > 0 {
+            if !l.is_empty() {
                 vec.push(l);
             }
         } else {
@@ -50,15 +50,15 @@ pub fn kmerize_string_set(l: &str, k: usize) -> fnv::FnvHashSet<std::string::Str
     let length_l = l.len();
     let l_r = revcomp(&l);
     if length_l < k {
-        set
+        return set;
     } else {
         for i in 0..l.len() - k + 1 {
             if seq::has_no_n(l[i..i + k].as_bytes()) {
                 set.insert(min(&l[i..i + k], &l_r[length_l - (i + k)..length_l - i]).to_string());
             }
         }
-        set
     }
+    set
 }
 
 #[inline]
@@ -68,7 +68,7 @@ pub fn kmerize_string_set_zeroth(l: &str, k: usize) -> fnv::FnvHashSet<u64> {
     let length_l = l.len();
     let l_r = revcomp(&l);
     if length_l < k {
-        set
+        return set;
     } else {
         for i in 0..l.len() - k + 1 {
             if seq::has_no_n(l[i..i + k].as_bytes()) {
@@ -82,8 +82,8 @@ pub fn kmerize_string_set_zeroth(l: &str, k: usize) -> fnv::FnvHashSet<u64> {
                 }
             }
         }
-        set
     }
+    set
 }
 
 //from https://docs.rs/bio/0.30.0/src/bio/alphabets/dna.rs.html#52-54
@@ -117,7 +117,7 @@ pub fn sliding_window_minimizers(
 ) -> fnv::FnvHashSet<std::string::String> {
     let mut window: VecDeque<(&str, usize)> = VecDeque::new();
     let mut mset = std::collections::HashSet::default();
-    let r_seq = revcomp(&seq);
+    let r_seq = revcomp(seq);
     let length = seq.len();
     for i in 0..seq.len() - m + 1 {
         let minimizer = min(&seq[i..i + m], &r_seq[length - (i + m)..length - i]);
@@ -128,7 +128,7 @@ pub fn sliding_window_minimizers(
         while (window.front().unwrap().1 as isize) < (i as isize - k as isize + m as isize) {
             window.pop_front();
         }
-        if i >= k - m && seq::has_no_n(&seq[i - (k - m)..i + m].as_bytes()) {
+        if i >= k - m && seq::has_no_n(seq[i - (k - m)..i + m].as_bytes()) {
             mset.insert(window.front().unwrap().0.to_string().to_uppercase());
         }
     }
@@ -136,7 +136,7 @@ pub fn sliding_window_minimizers(
 }
 
 pub fn sliding_window_minimizersi_zeroth(
-    v: &Vec<String>,
+    v: &[String],
     k: usize,
     m: usize,
 ) -> std::collections::HashSet<u64> {
@@ -146,7 +146,7 @@ pub fn sliding_window_minimizersi_zeroth(
             continue;
         }
         let mut window: VecDeque<(&str, usize)> = VecDeque::new();
-        let r_seq = revcomp(&seq);
+        let r_seq = revcomp(seq);
         let length = seq.len();
         for i in 0..seq.len() - m + 1 {
[…]
