+++ src/classify_batch.rs
@@ -1,18 +1,14 @@
 use super::build_index;
 use super::search_bits;
+use super::build_index::Parameters;
 use boomphf::*;
-use hashbrown::HashMap;
 
 #[allow(unused_assignments)]
 pub fn batch_classify(
     batch_samples: &str,
     db: &[u32],
-    taxonomy: &HashMap<u32, String>,
     phf: &Mphf<u64>,
-    lineage_graph: &HashMap<u32, u32>,
-    k: usize,
-    m: usize, //0 == no m, otherwise minimizer
-    value_bits: u32,
+    parameters: &Parameters,
     b: usize, //batch size for multi-threading
     qual_offset: u8,
     tag: &str,
@@ -30,13 +26,9 @@ pub fn batch_classify(
             if fq.len() > 1 {
                 search_bits::per_read_stream_pe(
                     &fq,
-                    &db,
-                    &taxonomy,
-                    &phf,
-                    &lineage_graph,
-                    k,
-                    m, //0 == no m, otherwise minimizer
-                    value_bits,
+                    db,
+                    phf,
+                    parameters,
                     b,
                     &prefix,
                     qual_offset, // q cutoff
@@ -45,13 +37,13 @@ pub fn batch_classify(
             } else {
                 search_bits::per_read_stream_se(
                     &fq,
-                    &db,
-                    &taxonomy,
-                    &phf,
-                    &lineage_graph,
-                    k,
-                    m, //0 == no m, otherwise minimizer
-                    value_bits,
+                    db,
+                    &parameters.taxonomy,
+                    phf,
+                    &parameters.lineage_graph,
+                    parameters.k_size,
+                    parameters.m_size, //0 == no m, otherwise minimizer
+                    parameters.value_bits,
                     b,
                     &prefix,
                     qual_offset, // q cutoff
@@ -61,13 +53,13 @@ pub fn batch_classify(
         } else {
             search_bits::per_read_stream_se(
                 &fq,
-                &db,
-                &taxonomy,
-                &phf,
-                &lineage_graph,
-                k,
-                m, //0 == no m, otherwise minimizer
-                value_bits,
+                db,
+                &parameters.taxonomy,
+                phf,
+                &parameters.lineage_graph,
+                parameters.k_size,
+                parameters.m_size, //0 == no m, otherwise minimizer
+                parameters.value_bits,
                 b,
                 &prefix,
                 0, // 0 for fasta or no quality filtering

[…]
