+++ src/main.rs
@@ -24,7 +24,7 @@ static GLOBAL: System = System;
 
 fn main() {
     let matches = App::new("sepia")
-            .version("0.0.2.")
+            .version("0.0.3.")
             .author("Henk C. den Bakker <henkcdenbakker@gmail.com>")
             .about("perfect hash index based read classifier and more")
             .setting(AppSettings::ArgRequiredElseHelp)
