### 2021-12-01 afda47d
Create rust.yml
Check out Github CI
