### 2021-12-01 3bfb0f6
Update rust.yml
Figuring out Github CI
