+++ .github/workflows/rust.yml
@@ -20,4 +20,5 @@ jobs:
       run: cargo build --verbose
     - name: Run tests
       run: cargo test --verbose
+    - name: Run sepia test
       run: bash test.sh
