### 2021-12-10 #3839 @hcdenbakker
Hi @Ebedthan

Thanks again for your quick review and helpful comments!

I believe the code will benefit (issue here) from incorporating some advises as suggested by cargo clippy.

- We changed the code to accommodate most of the suggestions provided by `cargo clippy`

The README is well written however it might benefit from a clearer installation section. The installation section should be divided into "precompiled binaries" and "install from source" sections. The release pages should be linked in the precompiled binaries section. The example code in the "install from source" section would be better on new lines like below:

        Clone the sepia repository

    git clone https://github.com/hcdenbakker/sepia.git

    The "Running sepia" section can be clearer by beginning with the "Build an index section" followed with "Classifying reads" and "Batch mode" thus showing the typical pipeline a user would follow. I think that the global sepia help page can come at the end of this section.

- We changed the README.MD as suggested by the reviewer

Using cargo test I've found only one unit test for the entire software. I recommend adding more tests and making the coverage of the test suite available in the README. Ideally, you should also set up continuous integration as
well which can be done with GitHub actions. Although I do not believe this is a specific requirement for JOSS I would advise it. You can find some examples on how to do it with bio crates like yacrd and sabreur.

- We switched the continuous integration from Travis CI to the Github based CI as suggested. We recognize the value of well written unit tests and plan to include unit tests in a future release of Sepia.

The To do at the end of the README can be placed inside the main.rs as a comment.

- This is an excellent suggestion, we will do this with future "To Do items"

The precompiled binaries are marked with a pre-release badge. If it is not intended, I would recommend removing it.

- The newest precompiled binaries are not marked with a pre-release badge anymore

I would suggest the release to have a proper naming strategy.

- We will follow a semantic versioning strategy from now on. The current release will be release v1.0.0

The precompiled binaries should contain the target triple.

- We made sure the target triple is in the precompiled binaries

In terms of the paper:

    Line 8: reword "the index"
    Line 26: add a comma before "and"
    Line 36-37: the indefinite article "a" may not be required with the plural noun "strings" in this sentence.

- We made the edits to the paper
