### 2021-12-11 #3839 @Ebedthan
Hi @hcdenbakker,

Thank you for the nice improvements made to the software, repository, and paper.
I only have one recommendation left, it is the adding of a community guideline in the README of the repository to guide people on how to contribute, seek help, or report issues. Making it clear can really be helpful especially for beginners.

Except for this last addition, I am good and recommend @will-rowe to accept the paper for publication.

Thanks again for your nice contribution to the field @hcdenbakker.
