### 2021-12-11 #3839 @telatin
Thanks @hcdenbakker for the changes, for the paper I'm satisfied!
I also recommend to accept the paper for publication.

I recently used Sepia and have some other comments that I will open as separate issues in the repo.
I like this addition to the acquarium ;)
