### 2021-12-16 #3839 @will-rowe
Hi @hcdenbakker!
This is looking good - I've just had another read and think we can move towards acceptance. We just need you to check your references and see if you can add DOIs for them (suggestions above). Once we have these done, we will need you to tag a new release and archive it. You then need to let me know the version and DOI for the archived release.

Thanks again reviewers - excellent reviews.
