### 2021-12-16 904ddd6
add DOIs to references
