+++ paper/paper.bib
@@ -7,6 +7,7 @@
   pages={1--13},
   year={2019},
   publisher={Springer}
+  doi={10.1186/s13059-019-1891-0}
 }
 
 @article{wood2014kraken,
@@ -18,6 +19,7 @@
   pages={1--12},
   year={2014},
   publisher={BioMed Central}
+  doi={10.1186/gb-2014-15-3-r46}
 }
 
 @article{kim2016centrifuge,
@@ -29,6 +31,7 @@
   pages={1721--1729},
   year={2016},
   publisher={Cold Spring Harbor Lab}
+  doi={10.1101/gr.210641.116}
 }
 
 @article{limasset2017fast,
@@ -47,6 +50,7 @@
   pages={996--1004},
   year={2018},
   publisher={Nature Publishing Group}
+  doi={10.1038/nbt.4229}
 }
 
 @misc{appleby2011murmurhash3,
@@ -64,6 +68,7 @@
   pages={1612--1625},
   year={2016},
   publisher={Cold Spring Harbor Lab}
+  doi={10.1101/gr.201863.115}
 }
 
 @article{segata2012metagenomic,
@@ -75,4 +80,5 @@
   pages={811--814},
   year={2012},
   publisher={Nature Publishing Group}
+  doi={10.1038/nmeth.2066}
 }
