### 2021-12-16 #3839 @hcdenbakker
Hi @will-rowe,
I have added the DOIs to the references and archived a new release (v1.0.1) at Zenodo (DOI 10.5281/zenodo.5787248)

Let me know if I need to do anything else,

Thanks,
-Henk
