### 2021-12-20 #3839 @telatin
>I only have one recommendation left, it is the adding of a community guideline in the README of the repository to guide people on how to contribute, seek help, or report issues. Making it clear can really be helpful especially for beginners.

I noticed that @Ebedthan also suggested improving the community guidelines (see quote).
I'm happy with the "Contributing" file you added in the repo root, from my side.
