### 2021-12-20 #3839 @Ebedthan
Hi @telatin,
Thank for the comment. Sorry @hcdenbakker I haven't seen the contributing file in the repo. I'm good with that and do not require it to be mentioned in the README, especially with the new github features which add a button to find it it easily.
Thanks.
