### 2021-12-22 #3839 @will-rowe
Sorry for the delay in response.

I'm afraid you need to update the zenodo archive @hcdenbakker as the author list is not correct. Please have the names appear as in your paper.
