### 2021-12-22 #3839 @hcdenbakker
Hi @will-rowe,

Sorry about that! I fixed the author names in the zenodo archive.

is there anything else I need to do?

Thanks
-Henk
