+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={1},
   pages={1--13},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1186/s13059-019-1891-0}
 }
 
@@ -18,7 +18,7 @@
   number={3},
   pages={1--12},
   year={2014},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
   doi={10.1186/gb-2014-15-3-r46}
 }
 
@@ -30,7 +30,7 @@
   number={12},
   pages={1721--1729},
   year={2016},
-  publisher={Cold Spring Harbor Lab}
+  publisher={Cold Spring Harbor Lab},
   doi={10.1101/gr.210641.116}
 }
 
@@ -49,7 +49,7 @@
   number={10},
   pages={996--1004},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   doi={10.1038/nbt.4229}
 }
 
@@ -67,7 +67,7 @@
   number={11},
   pages={1612--1625},
   year={2016},
-  publisher={Cold Spring Harbor Lab}
+  publisher={Cold Spring Harbor Lab},
   doi={10.1101/gr.201863.115}
 }
 
@@ -79,6 +79,6 @@
   number={8},
   pages={811--814},
   year={2012},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   doi={10.1038/nmeth.2066}
 }
