### 2021-12-22 #3839 @hcdenbakker
Hi @will-rowe,

OK, I should have caught the missing commas in the bibtex file. It is fixed now.

Thanks,
-Henk
