### 2021-12-22 #3839 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
